Add spelled_out_numbers_map, scrape deaths for LU
This change adds an associative array to map spelled out numbers to
their respective value. Using this we can detect the numbers of deaths
for the canton Lucerne.
It is backwords compatible. If there are no valid numbers found the
output of field `Deaths` is omitted.

Thanks to @baryluk who pointed me to the right files.

Closes openZH/covid_19#195