Add a matrix of expectations for each scraper in the parser.
This is an independent list that is checked against what scraper
provides. If any expected number is missing, it is reported as warning to
stderr.

In the future we might turn it into an error and fail parsing, to
indicate stronger that the scraper is broken or something.

This is related to https://github.com/openZH/covid_19/issues/68
and https://github.com/openZH/covid_19/issues/68#issuecomment-605527035
in particular.