Off-by-one when calling split in scraper parser. Add new date format for SH.
`"Date and time: 08.04.2020: 09.30 Uhr".split(": ", 2)` actually splits
into 3 components array / tuple, not 2.

'2' indicates not a (maximum) number of elements in the split, but number
of the split points. We want just 1.

This fixes parsing of todays SH scraping.

Similarly update calls to `"".split` in other places.

Also adjust one of the date regexps to capture a new date format.