Various small changes in parser and expectation matrix code.
Use `=` not `startswith`, with only few exceptions.

This is to allow extended numbers parsing, like maybe 'Confirmed cases
(non-residents)' separately.

Fix output of parsing '2020-03-01' to be '2020-03-01T'.

Match a dot in the scrape timestamp, for parts of a second, in the CSV
converter.

Reduce redundancy when parsing extras in the CSV convert.

Use `extras_dict` directly to pass extras to `check_expectations`.

Check for expected time of a day.

Enable 'SH' to have time in the expectation matrix. This was because of
recent commit to fix parsing of the date/time format of SH.

Print warning to stderr if one of the values from the scraper is
literally string `'None'`.