Omit url parameter for PDF source
FL adds a timestamp paramater (`ts`) to all uploads, this is done to
avoid delivering a cached version of the file (i.e. the browser is
forced to always download the latest file).
For the scraper this leads to the situation, that the source URL changed
with every scraper run, so that the File is always updated. This leads
to a lot of unneccessary commits.

This change omits all URL parameters, since they are not needed to get
the file.